requirements.txt
kagglehub
pandas
python-dotenv